Finished basic wizard

+++ app/controllers/select_timeslots_controller.rb
@@ -1,42 +1,53 @@
 class SelectTimeslotsController < ApplicationController
   include Wicked::Wizard
-  steps :monday, :tuesday, :wednesday, :thursday, :friday
+  steps :monday, :tuesday, :wednesday, :thursday, :friday, :rank, :summary
   
   def show
     @student = Student.find(params[:student_id])
+    case step
+    when :rank
+      @preferences = Preference.where(:student_id => @student.id)
+    end
     render_wizard 
   end
 
   def update
     student_id = params[:student_id]
     @student = Student.find(student_id)
-
-    # delte old preferences that were not selected again
-    delete_list = []
-    Preference.where(:student_id => student_id) do |p|
-      if not p.timeslot.day == step
-        delete_list
+    if step == :rank
+      params[:pref].each do |k, v|
+        p = Preference.find_by_id(k)
+        p.ranking = v
+        p.save!
       end
-    end
-    Preference.destroy_all(delete_list)
-
-
-    # create or update new preferences
-    now = DateTime.now
-
-    if params[step]
-      params[step].each do |timeslot_id|
-        p = Preference.where(:student_id => student_id, :id => timeslot_id).first || Preference.new()
-        p = p.update_attributes(
-          :student_id => student_id,
-          :timeslot_id => timeslot_id,
-          :updated_at => now,
-        )
-
+    else
+      # delete old preferences that were not selected again
+      delete_list = []
+      Preference.where(:student_id => student_id) do |p|
+        if not p.timeslot.day == step
+          delete_list
+        end
       end
-    end
-    
+      Preference.destroy_all(delete_list)
+
+
+      # create or update new preferences
+      now = DateTime.now
+
+      if params[step]
+        params[step].each do |timeslot_id|
+          p = Preference.where(:student_id => student_id, :id => timeslot_id).first || Preference.new()
+          p = p.update_attributes(
+            :student_id => student_id,
+            :timeslot_id => timeslot_id,
+            :updated_at => now,
+          )
+
+        end
+      end
+    end  
 
     render_wizard @student
   end
+
 end
